% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linen.R
\name{cells}
\alias{cells}
\title{Create cell contents}
\usage{
cells(ref, style, value, formula, is_formula, is_value, is_blank, is_bool,
  is_number, is_text, is_date)
}
\arguments{
\item{ref}{A cell reference in A1 format}

\item{style}{An integer indicating which style to apply}

\item{value}{A \emph{list} of values (NULL values when blank)}

\item{formula}{A \emph{list} of formulae (NULL values when blank)}

\item{is_formula}{Logical vector indicating if the cell is a formula}

\item{is_value}{Logical vector indicating if the cell is a value}

\item{is_blank}{Logical vector indicating if the cell is blank}

\item{is_bool}{Logical vector indicating if the cell is a boolean (logical)}

\item{is_number}{Logical vector indicating if the cell is a number (there is no distinction between integer and double)}

\item{is_text}{Logical vector indicating if the cell is text}

\item{is_date}{Logical vector indicating if the cell is a date}
}
\description{
Create a \code{tbl_df} of cell contents
}

