% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{linen_style}
\alias{linen_style}
\title{Something about style look up}
\usage{
linen_style(lookup, ...)
}
\arguments{
\item{lookup}{a parameter of the function}

\item{...}{further arguments passed to or from other methods}
}
\value{
a list
}
\description{
Jenny is just documenting this minimally because it needs to be exported to
please rexcel and for linen to pass R CMD check. She doesn't know enough yet
to write good docs.
}

