% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{style_lookup}
\alias{style_lookup}
\title{Look up and compute styles}
\usage{
style_lookup(x, ..., idx = NULL)
}
\arguments{
\item{x}{An object that can can converted into a
\code{linen_style} object; either the style data itself, or a
worksheet or workbook object.}

\item{...}{A set of elements to be extracted.  These must be
nested with "/" (e.g., "num_fmt/is_date_time"), and can be named
(in which case the output is named).}

\item{idx}{optional vector to subset the computed style with}
}
\description{
Work with \code{linen_style} objects.  This is going to need
documentation, but the details are still being worked out.
}
\author{
Rich FitzJohn
}

