% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{worksheet_to_table}
\alias{worksheet_to_table}
\title{Convert a linen worksheet to a data frame}
\usage{
worksheet_to_table(dat, col_names = TRUE, col_types = NULL, na = "",
  skip = 0)
}
\arguments{
\item{dat}{A linen \code{\link{worksheet}} object}

\item{col_names}{TRUE (the default) indicating we should use the
first row as column names, FALSE, indicating we should generate
names (X1, X2, ..., Xn) or a character vector of names to apply.}

\item{col_types}{Either NULL (the default) indicating we should
guess the column types or a vector of column types (values must
be "blank", "numeric", "date" or "text").}

\item{na}{Values indicating missing values (if different from
blank).  Not yet used.}

\item{skip}{Number of rows to skip.}

\item{sheet}{Sheet name or an integer}
}
\description{
Read a worksheet (from Excel or googlehseetws) the same way as
readxl, but slower.  Assumes a well behaved table of data.
}

