% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.R
\name{worksheet_view}
\alias{worksheet_view}
\title{Create worksheet view}
\usage{
worksheet_view(sheet, xr)
}
\arguments{
\item{sheet}{A worksheet object, created by
\code{\link{worksheet}} (possibly via the \code{googlesheets} or
\code{rexcel} packages).}

\item{xr}{A \code{cell_limits} object describing the range of the
worksheet to view.}
}
\description{
Create worksheet "view"
}

