% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linen.R
\name{cells}
\alias{cells}
\title{Create cell contents}
\usage{
cells(ref, style, value, formula, type)
}
\arguments{
\item{ref}{A cell reference in A1 format}

\item{style}{An integer indicating which style to apply}

\item{value}{A \emph{list} of values (NULL values when blank)}

\item{formula}{A \emph{list} of formulae (NULL values when blank)}

\item{type}{String describing the type of the cell.  Must be one
of "blank", "bool", "date", "number" or "text".}
}
\description{
Create a \code{tbl_df} of cell contents
}

