% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linen.R
\name{workbook}
\alias{workbook}
\title{Create a workbook object}
\usage{
workbook(names, style)
}
\arguments{
\item{names}{Names of sheets within the workbook}

\item{style}{A data.frame of style information.  This is not yet
validated and we'll change what goes on here at some point.}
}
\description{
Create a workbook object
}

