% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linen.R
\name{worksheet}
\alias{worksheet}
\title{Create a worksheet object}
\usage{
worksheet(name, cells, merged, workbook)
}
\arguments{
\item{name}{The name of the worksheet}

\item{cells}{A \code{tbl_df}, created by \code{\link{cells}}.}

\item{merged}{A list of \code{cell_limits} objects indicating
merged cellls.  Can be an empty list if there are no merged
cells.}

\item{workbook}{A workbook object.}
}
\description{
Create a worksheet object
}

